Update sublime theme template, add missing dd linenr
The sublime theme template had leftover references to old placeholder
values, which have been updated to their newer counterparts. The
template was also missing a setting for line number background and
foreground (gutter and gutterForeground), which are now added.

Also fixed the missing setting for line number color in the dusty
dunes theme.