Support percentage based RGB templates [skip ci]
For iTerm2, the existing color schemes can't be directly inserted into a
template, since iTerm2 requires RGB percentages.

This introduces a new method for translating a hex color code -- for
example #ffffff -- into a table of RGB percentages:
{
    R: 1.0,
    G: 1.0,
    B: 1.0
}
which can be used to fill out a template, as long as the template
variable is prefixed with its corresponding color (so instead of just
"color_fg", the value used in percentage based templates would be
"Rcolor_fg").