Rewrite theme generator in Zig
The theme generator portion of earthbound-themes has been extracted from
the original project and rewritten in Zig, with the intention of
distributing the generator as a CLI for others to use. The project has
always been built on a heavily templatized approach to allow dynamic
adding/updating of theme color values without having to manually go
through and edit each theme for each editor/term. This refactor takes
all of that work and extends it to allow others to use the same
functionality without being tied to creating themes that are only
relevant to Earthbound.

The refactor still needs a few polishing steps before releasing as a
package, but is fully functional in its current state.

As for the previous "Earthbound Themes" repo, my plan is to make that
into its own separate repo now, with Colorstorm as a dependency that it
can use to generate themes for all of the currently supported editors
and terminal emulators. On top of that, I'd like to start making more
SNES related theme packs, which Colorstorm should make a lot easier.

Zig also has the added bonus of being pretty stellar at cross
compilation, which should make distribution of the new tool much less of
a headache than I'd previously anticipated (hopefully).