Use new field for "safe" theme name
The previous field used for naming the theme was "theme_name_alt", which
didn't really make much sense. It's now been renamed to
"theme_name_safe", which encapsulates its purpose as a field that is
safe to be used as a filename without any potential conflicts (i.e.
names with spaces, forbidden chars, etc).