Initial rewrite of colorstorm to TUI [skip ci]
Colorstorm wasn't super useful as just a command line tool for taking
JSON and producing a set of theme files. Rewriting as a TUI allows the
user to see a preview of their color theme as they create it, along with
a reference image displayed (if one is provided).

This rewrite uses bubbletea for the TUI elements, as well as a custom
"mosaic" algorithm that converts a provided image into a set of ASCII
half block characters with their foreground and background colored as
sets of 2 pixels (top and bottom).