Use `gamut` for random theme gen, add mono theme gen [skip ci]
Switches to using the PastelPaletteGenerator from `gamut` for random
theme generation, which seems to look a lot better with a lot less logic
for modifying the palette after generation.

Adds a monochromatic theme generator, which can optionally take a
"seed" color to generate the theme. If no color is provided, it takes a
random color from gamut's HappyPaletteGenerator and uses that to
generate a monochromatic theme.

Gamut can also be used to determine if the foreground color in the form
color inputs should be white/black, rather than manually checking the
Hsv value of each color (which is what gamut is doing under the hood
anyways, but this reduces the amount of logic in form generation for
colorstorm).