Implement color picker [skip ci]
Adds a color picker to the TUI when a reference image is provided. This
works by tracking X and Y coordinates within the image and overlaying a
"cursor" over the drawn image pixels. The color beneath the cursor in
the color picker can be used for whichever theme color is currently
active (i.e. background, foreground, function, etc).

This performs bounds checking against min x/y (0) but not the max image
dimensions, meaning the cursor can leave the image. Not sure if that's
worth fixing since it should be pretty obvious to the user what happened
to the cursor.